+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Game Layers. ([304ee4c](https://github.com/NazioLT/NZCore/commit/304ee4cdee432c429569da9cfb8b2245b8fce5da))
 
 #### Editor
+* Custom NPropertyDrawer. ([]())
 * Display If. ([]())
 * Adapt GUI. ([]())